DRAFT: Working on tab's testing. Test with cassandra, almost done.

+++ twake/backend/node/src/services/channels/services/tab/service.ts
@@ -7,20 +7,22 @@ import {
   SaveResult,
   OperationType,
   UpdateResult,
+  CrudExeption,
 } from "../../../../core/platform/framework/api/crud-service";
-import { TabService } from "../../provider";
-import { ChannelTab, ChannelTabPrimaryKey } from "../../entities";
-import { ChannelExecutionContext } from "../../types";
+import {  MemberService, TabService } from "../../provider";
+import { ChannelMember, ChannelTab, ChannelTabPrimaryKey } from "../../entities";
+import { ChannelExecutionContext, ChannelVisibility } from "../../types";
 import { Channel } from "../../../types";
 import { DatabaseServiceAPI } from "../../../../core/platform/services/database/api";
 import { ResourcePath } from "../../../../core/platform/services/realtime/types";
 import Repository from "../../../../core/platform/services/database/services/orm/repository/repository";
 
+
 export class Service implements TabService {
   version: "1";
   repository: Repository<ChannelTab>;
 
-  constructor(private database: DatabaseServiceAPI) {}
+  constructor(private database: DatabaseServiceAPI, private members: MemberService) {}
 
   async init(): Promise<this> {
     this.repository = await this.database.getRepository("channel_tab", ChannelTab);
@@ -33,9 +35,9 @@ export class Service implements TabService {
       path: getTabsRealtimeResourcePath(tab, (context as ChannelExecutionContext).channel),
     },
   ])
-  async save<SaveOptions>(
+  async save<ChannelTabSaveOptions>(
     tab: ChannelTab,
-    options: SaveOptions,
+    options: ChannelTabSaveOptions,
     context: ChannelExecutionContext,
   ): Promise<SaveResult<ChannelTab>> {
     const pk = {
@@ -44,6 +46,22 @@ export class Service implements TabService {
       channel_id: context.channel.id,
       id: tab.id,
     };
+
+    const isDirectChannel = context.channel.workspace_id === ChannelVisibility.DIRECT;
+    if (isDirectChannel){
+      throw CrudExeption.badRequest("Driect channel dont have tabs");
+    }
+
+    const userMember = await this.members.get({
+      channel_id: pk.channel_id,
+      company_id: pk.company_id,
+      workspace_id: pk.workspace_id,
+      user_id: context.user.id,
+    });
+
+    if (!userMember) {
+      throw CrudExeption.badRequest("User is not channel member");
+    }
 
     let tabEntity = await this.repository.findOne(pk);
     if (!tabEntity) {
@@ -56,10 +74,12 @@ export class Service implements TabService {
       tabEntity,
       _.pick(tab, ["name", "configuration", "application_id", "order"]),
     );
-
+    
     await this.repository.save(tabEntity);
 
-    return new SaveResult("channel_tab", tabEntity, OperationType.CREATE);
+    const result =  new SaveResult("channel_tab", tabEntity, OperationType.CREATE);
+    this.onCreated(pk, tabEntity, result);
+    return result;
   }
 
   async get(tabPk: ChannelTabPrimaryKey, context: ChannelExecutionContext): Promise<ChannelTab> {
@@ -78,12 +98,22 @@ export class Service implements TabService {
       path: getTabsRealtimeResourcePath(tab, (context as ChannelExecutionContext).channel),
     },
   ])
-  async delete(pk: ChannelTabPrimaryKey): Promise<DeleteResult<ChannelTab>> {
+  async delete(pk: ChannelTabPrimaryKey, context: ChannelExecutionContext): Promise<DeleteResult<ChannelTab>> {
     let tabEntity = await this.repository.findOne(pk);
+    const userMember = await this.members.get({
+      channel_id: pk.channel_id,
+      company_id: pk.company_id,
+      workspace_id: pk.workspace_id,
+      user_id: context.user.id,
+    });
+    
+    if (!userMember) {
+      throw CrudExeption.badRequest("User is not channel member");
+    }
     if (tabEntity) {
       await this.repository.remove(tabEntity);
     }
-
+    this.onDeleted(pk, tabEntity);
     return new DeleteResult("channel_tab", tabEntity, true);
   }
 
@@ -121,6 +151,17 @@ export class Service implements TabService {
   onDeleted(channel: Channel, tab: ChannelTab): void {
     console.log("Tab deleted", tab);
   }
+
+  async userMember(pk: ChannelTabPrimaryKey, context: ChannelExecutionContext): Promise<ChannelMember>{
+    const member = await this.members.get({
+      channel_id: pk.channel_id,
+      company_id: pk.company_id,
+      workspace_id: pk.workspace_id,
+      user_id: context.user.id,
+    });
+    return member
+  }
+
 }
 
 export function getTabsRealtimeRoom(channel: ChannelExecutionContext["channel"]): string {
